requirements.txt
django
djangorestframework
python-dotenv
qdrant-client
langchain
sentence-transformers
PyMuPDF
pdf2image
pillow
boto3
openai
google-generativeai
anthropic
django_heroku
django-cors-headers
gunicorn
langchain_community
langchain
psycopg2